Bump publicodes to beta.37.
The current version is beta.66 but I can't make the code work with later
versions than beta.37.

- We now parse YAML code ourselves
- We add some EUR units (some tests with beta.38 could be fixed with
  this change, but not all of them)

Furthermore, with beta.44, publicodes seem to complain about multiple
definitions of the same variables, or loops in some definitions. It
seems that removing some branches in "variations" or "si/alors" removes
the complains (which I don't think makes sense), but not all of them.

I'll need more time to investigate (including maybe raising an issue on
GitHub).