Add robots, humans, security and sitemap files.
The humans.txt file is created with figlet.

The sitemap and robots.txt files are a first attempt: I'm not sure it's
a good idea to index every single gross to net computations (for each
possible amount, e.g. 3600.00, 3600.01, and so on, and for all variants,
e.g. with 1 child, 2 children, and so on). So instead, let's index only
two examples.